// frontend/wigo-travel/src/__generated__/PrismicSearchWidget.d.ts
/* tslint:disable */
/* eslint-disable */
// @generated
// This file was automatically generated and should not be edited.

// ====================================================
// GraphQL query operation: PrismicSearchWidget
// ====================================================

export interface PrismicSearchWidget_prismicSearchWidget_data_calendar_night_plural_label {
  text: string | null;
}

export interface PrismicSearchWidget_prismicSearchWidget_data_calendar_night_singular_label {
  text: string | null;
}

export interface PrismicSearchWidget_prismicSearchWidget_data_calendar_title {
  text: string | null;
}

export interface PrismicSearchWidget_prismicSearchWidget_data_cancel_label {
  text: string | null;
}

export interface PrismicSearchWidget_prismicSearchWidget_data_confirm_label {
  text: string | null;
}

export interface PrismicSearchWidget_prismicSearchWidget_data_destination_everywhere_label {
  text: string | null;
}

export interface PrismicSearchWidget_prismicSearchWidget_data_destinations_title {
  text: string | null;
}

export interface PrismicSearchWidget_prismicSearchWidget_data_people_plural_label {
  text: string | null;
}

export interface PrismicSearchWidget_prismicSearchWidget_data_people_singular_label {
  text: string | null;
}

export interface PrismicSearchWidget_prismicSearchWidget_data_people_title {
  text: string | null;
}

export interface PrismicSearchWidget_prismicSearchWidget_data_title {
  text: string | null;
}

export interface PrismicSearchWidget_prismicSearchWidget_data_api_error_message {
  text: string | null;
}

export interface PrismicSearchWidget_prismicSearchWidget_data {
  calendar_night_plural_label: (PrismicSearchWidget_prismicSearchWidget_data_calendar_night_plural_label | null)[] | null;
  calendar_night_singular_label: (PrismicSearchWidget_prismicSearchWidget_data_calendar_night_singular_label | null)[] | null;
  calendar_title: (PrismicSearchWidget_prismicSearchWidget_data_calendar_title | null)[] | null;
  cancel_label: (PrismicSearchWidget_prismicSearchWidget_data_cancel_label | null)[] | null;
  confirm_label: (PrismicSearchWidget_prismicSearchWidget_data_confirm_label | null)[] | null;
  destination_everywhere_label: (PrismicSearchWidget_prismicSearchWidget_data_destination_everywhere_label | null)[] | null;
  destinations_title: (PrismicSearchWidget_prismicSearchWidget_data_destinations_title | null)[] | null;
  people_plural_label: (PrismicSearchWidget_prismicSearchWidget_data_people_plural_label | null)[] | null;
  people_singular_label: (PrismicSearchWidget_prismicSearchWidget_data_people_singular_label | null)[] | null;
  people_title: (PrismicSearchWidget_prismicSearchWidget_data_people_title | null)[] | null;
  title: (PrismicSearchWidget_prismicSearchWidget_data_title | null)[] | null;
  api_error_message: (PrismicSearchWidget_prismicSearchWidget_data_api_error_message | null)[] | null;
}

export interface PrismicSearchWidget_prismicSearchWidget {
  data: PrismicSearchWidget_prismicSearchWidget_data | null;
}

export interface PrismicSearchWidget {
  prismicSearchWidget: PrismicSearchWidget_prismicSearchWidget | null;
}